refactor: GameStatus 패키지 이동, 검증 로직 의존관계 없애기
