+++ README.md
@@ -223,7 +223,7 @@ sequenceDiagram
     - [x] Piece.canMove~() 에서 이동방향 검증까지 직접 할 수 있도록 개선
         - Piece 로 체스판의 이동 검증 로직 책임 위임 및 통합
     - [ ] ActionMapper `Enum 요소 조회 비용` vs `Map 클래스로 변경 후, static 상수로 인한 메모리 사용` 비교
-    - [ ] GameStatus 패키지 이동, 검증 로직 의존관계 없애기
+    - [x] GameStatus 패키지 이동, 검증 로직 의존관계 없애기
     - [ ] King 공격 확인 메서드 분리
 - 코드 관련
     - [ ] Dao 템플릿 적용하여 중복 코드 제거
